# make function fields type stable