--- a/src/de.jl
+++ b/src/de.jl
@@ -10,16 +10,16 @@ The constructor takes following keyword arguments:
 - `recombination`: the recombination functions (default: [`uniformbin(0.5)`](@ref))
 - `K`: the recombination scale factor (default: 0.5*(F+1))
 """
-@kwdef struct DE <: AbstractOptimizer
+@kwdef struct DE{F1,F2} <: AbstractOptimizer
     populationSize::Integer = 50
     F::Real = 0.9
     n::Integer = 1
-    K::Real = 0.5*(F+1)
-    selection::Function = random
-    recombination::Function = uniformbin(0.5)
+    K::Real = 0.5 * (F + 1)
+    selection::F1 = random
+    recombination::F2 = uniformbin(0.5)
 end
 population_size(method::DE) = method.populationSize
-default_options(method::DE) = (abstol=1e-10,)
+default_options(method::DE) = (abstol = 1e-10,)
 summary(m::DE) = "DE/$(m.selection)/$(m.n)/$(m.recombination)"
 
 mutable struct DEState{T,IT} <: AbstractOptimizerState
@@ -60,7 +60,7 @@ function update_state!(objfun, constraints, state, population::AbstractVector{IT
         offspring[i] = copy(base)
         # println("$i => base:", offspring[i])
 
-        targets = randexcl(1:Np, [i], 2*n)
+        targets = randexcl(1:Np, [i], 2 * n)
         offspring[i] = differentiation(offspring[i], @view population[targets]; F=F)
         # println("$i => mutated:", offspring[i], ", targets:", targets)