I noticed that the fields in the structs were abstract for functions, leading to type instability. This PR makes the functions concrete. The benchmark provided in the package showed a moderate improvement. Of course, this will vary depending on the details of the setup. 
Before:
```
            "ga" => Trial(510.096 μs)
            "cmaes" => Trial(82.302 ms)
            "es" => Trial(1.185 ms)
            "de" => Trial(1.030 ms)
```

After: 
```
            "ga" => Trial(400.469 μs)
            "cmaes" => Trial(85.420 ms)
            "es" => Trial(1.154 ms)
            "de" => Trial(750.362 μs)
```
Note that no changes were made to cmaes.